Add files via upload
This is a predator-prey model that models the population of elk (prey) being hunted by wolves (predator) based on the reintroduction of wolves into the Yellowstone national park. Accompanied is the essay which I wrote for a project that this model is based on. Original model was designed in Maple and is remade using Simulink.